--- a/pfring/pfring_test.go
+++ b/pfring/pfring_test.go
@@ -1,3 +1,9 @@
+// Copyright 2019 The GoPacket Authors. All rights reserved.
+//
+// Use of this source code is governed by a BSD-style license
+// that can be found in the LICENSE file in the root of the source
+// tree.
+
 package pfring
 
 import (