Rework of the core parts of pfring. Due to the use of packed structures in
7.2 cgo can't access some fields of `struct pfring_pkthdr`.
This copies the needed fields to a non-packed struct first.

To improve speed, `pfring_recv` is now called with a length of zero,
enabling zero copy. The old version caused the userlab pf_ring library
to copy the packet contents. This maybe fixes #543, #185.
Additionally, pfring now uses `timestamp_ns` instead of `ts`, which increases
the timestamp resolution from microsecond to nanosecond. Furthermore,
cgo usage was improved by using the same techniques as in pcap
(allocations on the heap).

Added `ZeroCopyReadPacketData`. This one uses not zero copy from kernel
space directly to the buffer. The old `ReadPacketDataTo` caused pf_ring
userspace library to copy the data (buffer_len != 0). The new
`ReadPacketDataTo` still does the copy (but now in gocode) since zero
copy is not possible with the current interface. This method is there to
keep legacy code working, but newer versions should switch over to
`ZeroCopyReadPacketData`.

The following benchmarks were performed in a virtual machine by using
pktgen kernel module on the host and virtio-net in the guest. Since
there is no `ZeroCopyReadPacketData`, the `PfringReadZero` lines compare
`ReadPacketDataTo` in the old implementation with
`ZeroCopyReadPacketData` in the new version.

65536 byte size packets:
```
name              old time/op    new time/op    delta
PfringRead-2        47.5µs ±10%    42.8µs ± 7%    -9.85%  (p=0.012 n=8+10)
PfringReadZero-2    42.5µs ±25%    32.9µs ± 7%   -22.65%  (p=0.000 n=10+10)
PfringReadTo-2      42.5µs ±25%    36.8µs ± 9%      ~     (p=0.143 n=10+10)

name              old speed      new speed      delta
PfringRead-2      1.30GB/s ±26%  1.53GB/s ± 7%   +17.57%  (p=0.003 n=10+10)
PfringReadZero-2  1.58GB/s ±22%  1.99GB/s ± 7%   +25.90%  (p=0.000 n=10+10)
PfringReadTo-2    1.58GB/s ±22%  1.79GB/s ± 9%      ~     (p=0.143 n=10+10)

name              old alloc/op   new alloc/op   delta
PfringRead-2        65.6kB ± 0%    65.5kB ± 0%    -0.02%  (p=0.000 n=10+10)
PfringReadZero-2     16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
PfringReadTo-2       16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)

name              old allocs/op  new allocs/op  delta
PfringRead-2          2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=10+10)
PfringReadZero-2      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
PfringReadTo-2        1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
```

58 byte size packets:
```
name              old time/op    new time/op     delta
PfringRead-2        46.8µs ±23%      3.2µs ±85%    -93.15%  (p=0.000 n=10+10)
PfringReadZero-2    2.48µs ±11%     2.33µs ± 2%     -6.37%  (p=0.022 n=10+8)
PfringReadTo-2      2.48µs ±11%     2.63µs ± 6%       ~     (p=0.053 n=10+9)

name              old speed      new speed       delta
PfringRead-2      1.27MB/s ±24%  21.17MB/s ±60%  +1570.88%  (p=0.000 n=10+10)
PfringReadZero-2  23.5MB/s ±10%   24.9MB/s ± 2%     +6.33%  (p=0.017 n=10+8)
PfringReadTo-2    23.5MB/s ±10%   22.0MB/s ± 6%       ~     (p=0.053 n=10+9)

name              old alloc/op   new alloc/op    delta
PfringRead-2        65.6kB ± 0%      0.1kB ± 0%    -99.90%  (p=0.000 n=10+10)
PfringReadZero-2     16.0B ± 0%       0.0B        -100.00%  (p=0.000 n=10+10)
PfringReadTo-2       16.0B ± 0%       0.0B        -100.00%  (p=0.000 n=10+10)

name              old allocs/op  new allocs/op   delta
PfringRead-2          2.00 ± 0%       1.00 ± 0%    -50.00%  (p=0.000 n=10+10)
PfringReadZero-2      1.00 ± 0%       0.00        -100.00%  (p=0.000 n=10+10)
PfringReadTo-2        1.00 ± 0%       0.00        -100.00%  (p=0.000 n=10+10)
```