@notti 2018-11-01
@gconnell any idea how govet fails here? The pfring directory isn't checked by govet, I didn't change anything outside it, and if I execute go vet on my computer there are no errors.